neet-stickers/create_neet_spin.py
import json
import os
from PIL import Image
import numpy as np

def create_neet_spin_animation(svg_path, output_path):
    """
    Creates a spinning NEET APE animation with rainbow effects
    
    Args:
        svg_path (str): Path to the NEET APE SVG file
        output_path (str): Path to save the Lottie animation
    """
    # Load the base animation template
    with open('lottie_animations/neet_ape_spin.json', 'r') as f:
        animation_data = json.load(f)
    
    # Load and process SVG
    # Note: You'll need to implement SVG processing based on your specific needs
    # This is a placeholder for the actual SVG processing logic
    
    # Add the processed SVG data to the shapes array
    # animation_data['layers'][0]['shapes'] = processed_svg_shapes
    
    # Save the final animation
    with open(output_path, 'w') as f:
        json.dump(animation_data, f, indent=2)

def main():
    # Create necessary directories
    os.makedirs('lottie_animations', exist_ok=True)
    
    # Paths
    svg_path = 'neet_simple.svg'  # Using the simplified SVG version
    output_path = 'lottie_animations/neet_ape_spin_complete.json'
    
    # Create the animation
    create_neet_spin_animation(svg_path, output_path)
    print(f"Animation created and saved to {output_path}")

if __name__ == "__main__":
    main()
