neet-stickers/requirements.txt
python-telegram-bot==20.6
Pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0.74
rembg>=2.0.50
torch>=2.0.1
face-recognition>=1.3.0
mediapipe>=0.10.3
dlib>=19.24.2
requests==2.31.0
moviepy==1.0.3
selenium==4.11.2
webdriver-manager==4.0.0
